+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+* --filter-rh-naming enabled by default (GRIF-121)
+
+### Added
+ * --filter-rh-naming default value can be set via .griffonrc 
+   in default section(GRIG-121)
+
 
 ## [0.5.0] - 2024-01-18